Adding MySQL Integration Test Setup and MySQL Connection Service Integration Test
