+++ requirements.txt
@@ -21,3 +21,4 @@ flake8==3.7.0
 ptvsd==4.3.2
 python-dateutil
 pydevd
+cryptography
