+++ tests/integration/connections_for_integration_tests.py
@@ -8,6 +8,7 @@
 from abc import abstractclassmethod
 from typing import List
 import psycopg2
+import pymysql
 
 
 class IConnection():
@@ -47,6 +48,26 @@ class PostgreSQLConnection:
         cls._connections = []
         for config_dict in connection_details_list:
             connection = psycopg2.connect(**config_dict)
+            cls._connections.append(connection)
+            connection.autocommit = True
+
+    @classmethod
+    def get_connections(cls):
+        return cls._connections
+
+    @classmethod
+    def get_connection_server_version(cls, index):
+        return cls._connections[index].server_version
+
+class MySQLConnection:
+    
+    _connections: List[pymysql.connections.Connection] = []
+
+    @classmethod
+    def open_connections(cls, connection_details_list):
+        cls._connections = []
+        for config_dict in connection_details_list:
+            connection = pymysql.connect(**config_dict)
             cls._connections.append(connection)
             connection.autocommit = True
