+++ tests/json_rpc_tests/rpc_test_framework.py
@@ -10,6 +10,7 @@ import json
 import logging
 import os
 import re
+import sys
 import threading
 from typing import Callable, List, Optional, Tuple
 from unittest import mock
@@ -85,9 +86,9 @@ class JSONRPCTestCase:
         shutdown_messages = [DefaultRPCTestMessages.shutdown()]
         self.messages = initialization_messages + test_messages + shutdown_messages
 
-    def run(self):
+    def run(self, provider=constants.PG_PROVIDER_NAME):
         # Start the server
-        input_stream, output_stream, output_info = JSONRPCTestCase.start_service()
+        input_stream, output_stream, output_info = JSONRPCTestCase.start_service(provider)
         output = ""
 
         # Send all messages to the server
@@ -172,7 +173,7 @@ class JSONRPCTestCase:
                 [json.dumps(notifications[index]) for index in notifications_to_verify])))
 
     @staticmethod
-    def start_service():
+    def start_service(provider=constants.PG_PROVIDER_NAME):
         # Set up the server's input and output
         input_r, input_w = os.pipe()
         server_input_stream = open(input_r, 'rb', buffering=0, closefd=False)
@@ -195,7 +196,7 @@ class JSONRPCTestCase:
 
         logger = logging.Logger('test')
         logger.addHandler(logging.NullHandler())
-        server = ossdbtoolsservice_main._create_server(server_input_stream, server_output_stream, logger, constants.PG_PROVIDER_NAME)
+        server = ossdbtoolsservice_main._create_server(server_input_stream, server_output_stream, logger, provider)
         server.start()
         return test_input_stream, server_output_stream, output_info
 
@@ -217,7 +218,7 @@ class DefaultRPCTestMessages:
     def change_configuration():
         return RPCTestMessage(
             'workspace/didChangeConfiguration',
-            '{"settings":{"pgsql":{"logDebugInfo":false,"enabled":true,"defaultDatabase":"postgres","format":{"keywordCase":null,"identifierCase":null,"stripComments":false,"reindent":true}}}}',  # noqa
+            '{"settings":{"pgsql":{"logDebugInfo":false,"enabled":true,"defaultDatabase":"postgres","format":{"keywordCase":null,"identifierCase":null,"stripComments":false,"reindent":true}},"MySQL":{"logDebugInfo":false,"enabled":true,"defaultDatabase":"","format":{"keywordCase":null,"identifierCase":null,"stripComments":false,"reindent":true}}}}',  # noqa
             JSONRPCMessageType.Notification
         )
 
@@ -230,7 +231,7 @@ class DefaultRPCTestMessages:
         )
 
     @staticmethod
-    def connection_request(owner_uri, connection_options):
+    def connection_request(owner_uri, connection_options, provider=constants.PG_PROVIDER_NAME):
         connection_request = RPCTestMessage(
             'connection/connect',
             '{"ownerUri":"%s","connection":{"options":%s}}' % (owner_uri, json.dumps(connection_options)),
@@ -242,7 +243,7 @@ class DefaultRPCTestMessages:
         )
         language_flavor_notification = RPCTestMessage(
             'connection/languageflavorchanged',
-            '{"uri":"%s","language":"sql","flavor":"PGSQL"}' % owner_uri,
+            '{"uri":"%s","language":"sql","flavor":"%s"}' % (owner_uri, provider),
             JSONRPCMessageType.Notification,
             notification_verifiers=[(
                 lambda notification: notification['method'] == 'textDocument/intelliSenseReady' and notification['params']['ownerUri'] == owner_uri,
